requirements.txt
mysqlclient~=2.1.1
pandas~=1.5.3
Flask~=2.2.3
Werkzeug~=2.2.3
numpy~=1.23.5
python-dateutil~=2.8.2
keras~=2.12.0
sklearn~=0.0.post4
scikit-learn~=1.2.2